Added GetAddress property to check if the address exist

+++ AdeCartAPI/Controllers/AddressController.cs
@@ -2,12 +2,8 @@
 using AdeCartAPI.UserModel;
 using AdeCartAPI.UserService;
 using AutoMapper;
-using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Threading.Tasks;
 
 namespace AdeCartAPI.Controllers
@@ -26,7 +22,7 @@ namespace AdeCartAPI.Controllers
             this.mapper = mapper;
         }
         [HttpPost]
-        public async Task<ActionResult> Post(string username,AddressCreate userAddress)
+        public async Task<ActionResult> Post(string username, AddressCreate userAddress)
         {
             var currentUser = await GetUser(username);
             if (currentUser == null) return NotFound();
@@ -34,26 +30,30 @@ namespace AdeCartAPI.Controllers
             var userId = mapper.Map<UserAddress>(currentUser);
             newAddress.UserId = userId.UserId;
             await address.CreateAddress(newAddress);
-            return Ok();
+            return Ok("Address successfully added");
         }
         [HttpPut]
-        public async Task<ActionResult>Put(string username,AddressUpdate userAddress) 
+        public async Task<ActionResult> Put(string username, AddressUpdate userAddress)
         {
             var currentUser = await GetUser(username);
             if (currentUser == null) return NotFound();
-            var newAddress = mapper.Map<UserAddress>(userAddress);
+            var updateAddress = mapper.Map<UserAddress>(userAddress);
+            var isExist = address.GetAddress(updateAddress.AddressId);
+            if (isExist == 0) return NotFound("The address doesn't exist");
             var userId = mapper.Map<UserAddress>(currentUser);
-            newAddress.UserId = userId.UserId;
-            await address.UpdateAddress(newAddress);
+            updateAddress.UserId = userId.UserId;
+            await address.UpdateAddress(updateAddress);
             return Ok();
         }
-        [HttpDelete]
-        public async Task<ActionResult> Delete(string username,Guid id) 
+        [HttpDelete("{id}")]
+        public async Task<ActionResult> Delete(string username,int id) 
         {
             var currentUser = await GetUser(username);
             if (currentUser == null) return NotFound();
+            var isExist = address.GetAddress(id);
+            if (isExist == 0) return NotFound("The address doesn't exist");
             await address.DeleteAddress(id);
-            return Ok();
+            return NoContent();
         }
         private async Task<User> GetUser(string userName)
         {
